requirements.txt
pycaw==20251023
keyboard==0.13.5
comtypes==1.4.14
psutil==7.2.1
pillow==12.1.0
pywin32==311
pystray==0.19.5
PySide6>=6.6
